src/broadcast/requirements.txt
python-dotenv
telethon
qrcode[pil]
